update user home page
Memindahkan folder user-home-pages ke dalam folder user, kemudian mengganti nama foldernya menjadi home
